Structured the code into modules for readability
Also improved the printout of the solution.
